pyproject.toml
[tool.poetry]
name = "rocketsync"
version = "0.1.0"
description = "Sync and organize your rocketbook scans"
authors = ["Evann DREUMONT <53308142+LeGmask@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.39.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.0"
tqdm = "^4.63.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
